KeyboardAvoidingView  and extract EditableField
